+++ tests/geopage/migrations/0001_initial.py
@@ -24,6 +24,8 @@ class Migration(migrations.Migration):
         ("wagtailcore", "0066_collection_management_permissions"),
     ]
 
+    streamfield_params = {"use_json_field": True} if WAGTAIL_VERSION >= (3, 0) else {}
+
     operations = [
         migrations.CreateModel(
             name="ClassicGeoPage",
@@ -382,7 +384,8 @@ class Migration(migrations.Migration):
                                     icon="user",
                                 ),
                             ),
-                        ]
+                        ],
+                        **streamfield_params,
                     ),
                 ),
             ],