+++ tests/geopage_nospatial/models.py
@@ -2,14 +2,19 @@ from django.db import models
 from django.utils.functional import cached_property
 from django.utils.translation import gettext as _
 from wagtail import VERSION as WAGTAIL_VERSION
-from wagtail.admin.edit_handlers import FieldPanel, MultiFieldPanel, StreamFieldPanel
 
 if WAGTAIL_VERSION >= (3, 0):
     from wagtail import blocks
+    from wagtail.admin.panels import FieldPanel, MultiFieldPanel
     from wagtail.fields import StreamField
     from wagtail.models import Page
 else:
     from wagtail.core import blocks
+    from wagtail.admin.edit_handlers import (
+        FieldPanel,
+        MultiFieldPanel,
+        StreamFieldPanel,
+    )
     from wagtail.core.fields import StreamField
     from wagtail.core.models import Page
 
@@ -21,11 +26,7 @@ from wagtailgeowidget.blocks import (
     GoogleMapsBlock,
     LeafletBlock,
 )
-from wagtailgeowidget.edit_handlers import (
-    GeoAddressPanel,
-    GoogleMapsPanel,
-    LeafletPanel,
-)
+from wagtailgeowidget.panels import GeoAddressPanel, GoogleMapsPanel, LeafletPanel
 
 
 class StandardPage(Page):
@@ -175,6 +176,8 @@ class StandardPageWithLeafletAndZoom(Page):
 
 
 class StreamPage(Page):
+    streamfield_params = {"use_json_field": True} if WAGTAIL_VERSION >= (3, 0) else {}
+
     body = StreamField(
         [
             ("map", GoogleMapsBlock()),
@@ -240,9 +243,15 @@ class StreamPage(Page):
                     icon="user",
                 ),
             ),
-        ]
+        ],
+        **streamfield_params,
     )
 
-    content_panels = Page.content_panels + [
-        StreamFieldPanel("body"),
-    ]
+    if WAGTAIL_VERSION >= (3, 0):
+        content_panels = Page.content_panels + [
+            FieldPanel("body"),
+        ]
+    else:
+        content_panels = Page.content_panels + [
+            StreamFieldPanel("body"),
+        ]