+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## Unreleased
+### Added
+- Add Wagtail 4 compability (@katdom13)
+
+### Changed
+- Updated StreamFieldPanel to just FieldPanel in tests (@katdom13)
+- Updated StreamFields to have additional argument use_json_field
+- Renamed wagtailgeowidget.edit_handlers to wagtailgeowidget.panels
+- Updated imports in docs (@katdom13)
+
+
 ## [6.2.0] - 2022.07.03
 ### Added
 - Add Wagtail 3 compability (@marteinn)